% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robomit_functions.R
\name{o_beta_boot_inf}
\alias{o_beta_boot_inf}
\title{Bootstrapped mean beta* and confidence intervals}
\usage{
o_beta_boot_inf(y, x, con, m = "none", w = NULL, id = "none", time = "none",
delta = 1, R2max, sim, obs, rep, CI, type, useed = NA, data)
}
\arguments{
\item{y}{Name of the dependent variable (as string).}

\item{x}{Name of the independent treatment variable (i.e., variable of interest; as string).}

\item{con}{Name of related control variables. Provided as string in the format: "w + z +...".}

\item{m}{Name of unrelated control variables (m; see Oster 2019; as string; default is m = "none").}

\item{w}{weights (only for weighted estimations). Warning: For weighted panel models R can report different R-square than Stata, leading deviation between R and Stata results.}

\item{id}{Name of the individual id variable (e.g. firm or farm; as string). Only applicable for fixed effect panel models.}

\item{time}{Name of the time id variable (e.g. year or month; as string). Only applicable for fixed effect panel models.}

\item{delta}{delta for which beta*s should be estimated (default is delta = 1).}

\item{R2max}{Maximum R-square for which beta*s should be estimated.}

\item{sim}{Number of simulations.}

\item{obs}{Number of draws per simulation.}

\item{rep}{Bootstrapping either with (= TRUE) or without (= FALSE) replacement}

\item{CI}{Confidence intervals, indicated as vector. Can be and/or 90, 95, 99.}

\item{type}{Model type (either \emph{lm} or \emph{plm}; as string).}

\item{useed}{User defined seed.}

\item{data}{Dataset.}
}
\value{
Returns tibble object, which includes the mean and confidence intervals of estimated bootstrapped beta*s.
}
\description{
Provides the mean and confidence intervals of estimated bootstrapped beta*s, i.e., the bias-adjusted treatment effects (or correlations) (following Oster 2019).
}
\details{
Provides the mean and confidence intervals of estimated bootstrapped beta*s, i.e., the bias-adjusted treatment effects (or correlations) (following Oster 2019). Bootstrapping can either be done with or without replacement. The function supports linear cross-sectional (see \emph{lm} objects in R) and fixed effect panel (see \emph{plm} objects in R) models.
}
\examples{
# load data, e.g. the in-build mtcars dataset
data("mtcars")
data_oster <- mtcars

# preview of data
head(data_oster)

# load robomit
require(robomit)

# compute the mean and confidence intervals of estimated bootstrapped beta*s
o_beta_boot_inf(y = "mpg",            # dependent variable
                x = "wt",             # independent treatment variable
                con = "hp + qsec",    # related control variables
                delta = 1,            # delta
                R2max = 0.9,          # maximum R-square
                sim = 100,            # number of simulations
                obs = 30,             # draws per simulation
                rep = FALSE,          # bootstrapping with or without replacement
                CI = c(90,95,99),     # confidence intervals
                type = "lm",          # model type
                useed = 123,          # seed
                data = data_oster)    # dataset
}
\references{
Oster, E. (2019). Unobservable Selection and Coefficient Stability: Theory and Evidence. Journal of Business & Economic Statistics, 37, 187-204.
}
