% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robomit_functions.R
\name{o_delta_boot_viz}
\alias{o_delta_boot_viz}
\title{Visualization of bootstrapped deltas*}
\usage{
o_delta_boot_viz(y, x, con, id = "none", time = "none", beta = 0, R2max, sim, obs, rep,
CI, type, norm = TRUE, bin, col = c("#08306b","#4292c6","#c6dbef"),
nL = FALSE, mL = TRUE, useed = NA, data)
}
\arguments{
\item{y}{Name of the dependent variable (as string).}

\item{x}{Name of the independent variable of interest (treatment variable; as string).}

\item{con}{Name of the other control variables. Provided as string in the format: "w + z +...".}

\item{id}{Name of the individual id variable (e.g. firm or farm; as string). Only applicable for fixed effect models.}

\item{time}{Name of the time variable (e.g. year or month; as string). Only applicable for fixed effect models.}

\item{beta}{Beta for which delta* should be estimated (default is beta = 0).}

\item{R2max}{Max R-square for which beta* should be estimated.}

\item{sim}{Number of simulations.}

\item{obs}{Number of draws per simulation.}

\item{rep}{Bootstrapping either with (= TRUE) or without (= FALSE) replacement}

\item{CI}{Confidence intervals, indicated as vector. Can be and/or 90,95,99.}

\item{type}{Model type (either \emph{lm} or \emph{plm}; as string).}

\item{norm}{Option to include a normal distribution in the plot (default is norm = TURE).}

\item{bin}{Number of bins used for the histogram.}

\item{col}{Colors used to indicate different confidence interval levels (indicated as vector). Needs to be the same length as the variable CI. The default is a blue color range.}

\item{nL}{Option to include a red vertical line at 0 (default is nL = TRUE).}

\item{mL}{Option to include a vertical line at beta* mean (default is mL = TRUE).}

\item{useed}{Seed number defined by user.}

\item{data}{Data.}
}
\value{
Returns ggplot object. Including bootstrapped deltas*.
}
\description{
Estimates and visualizes bootstrapped deltas*, i.e. the degree of selection on unobservables relative to observables that would be necessary to explain away the result, following Oster (2019).
}
\details{
Estimates and visualizes bootstrapped deltas*, i.e. the degree of selection on unobservables relative to observables that would be necessary to explain away the result, following Oster (2019). Bootstrapping can either be done with or without replacement. The function supports linear cross sectional (see \emph{lm} objects in R) and panel fixed effect (see \emph{plm} objects in R) models.
}
\examples{
# load data, e.g. the in-build mtcars dataset
data("mtcars")
data_oster <- mtcars

# preview of data
head(data_oster)

# load robomit
require(robomit)

# estimate and visualize bootstrapped deltas*
o_delta_boot_viz(y = "mpg",               # define the dependent variable name
                     x = "wt",            # define the main independent variable name
                     con = "hp + qsec",   # other control variables
                     beta = 0,            # define beta. This is usually set to 0
                     R2max = 0.9,         # define the max R-square.
                     sim = 100,           # define number of simulations
                     obs = 30,            # define number of drawn observations per simulation
                     rep = FALSE,         # define if bootstrapping is with or without replacement
                     CI = c(90,95,99),    # define confidence intervals.
                     type = "lm",         # define model type
                     norm = TRUE,         # include normal distribution
                     bin = 200,           # set number of bins
                     useed = 123,         # define seed
                     data = data_oster)   # define dataset
}
\references{
Oster, E. (2019). Unobservable selection and coefficient stability: Theory and evidence. Journal of Business & Economic Statistics, 37, 187-204.
}
